+++ precompute.h
@@ -10,7 +10,7 @@ using namespace scene;
 /* Precompute Functions */
 void precomputeMeshNormals(const vector<Mesh>& meshes, vector<vector<VectorFloatTriplet>>& meshVertexNormals, const vector<VectorFloatTriplet>& vertices);
 void precomputeTriangleNormals(const vector<Triangle>& triangles, vector<VectorFloatTriplet>& triangleNormals, const vector<VectorFloatTriplet>& vertices);
-void precomputeCameraTriangleDeterminant(const Scene& scene, vector<vector<float>>& cameraTriangleDeterminant);
-void precomputeCameraMeshDeterminant(const Scene& scene, vector<vector<vector<float>>>& cameraMeshDeterminant);
+void precomputeCameraTriangleDeterminant(const Scene& scene, vector<vector<double>>& cameraTriangleDeterminant);
+void precomputeCameraMeshDeterminant(const Scene& scene, vector<vector<vector<double>>>& cameraMeshDeterminant);
 
 #endif
